Don't use bitcast to cast integers to the i1 type.
I can't seem to find an llvm instruction that does what I want with
casting. The only viable way seems to be to use the same type with the
comparison. Since right now we only support int's anyway, I've changed
it so it will create the appropriately sized int based on the other
operand.

Using bitcast doesn't work. Using truncate will make any even number
evaluate to 0. So instead we just compare with the proper size.