Basic function passes for optimizing the resulting LLVM IR code.
Right now I'm just using the CFG simplification pass which eliminates dead
blocks, merges basic blocks if possible, and some other optimizations.

I also modified how If generates instructions to be simpler. It was very
complicated (and didn't work when nesting if statements) because it was
trying to avoid creating dead blocks and avoid unnecessary branching.

Because of the function pass, this isn't needed and it simplifies the code
a lot to not worry about the IR code nearly as much.